+++ daisy/model/NFMRecommender.py
@@ -128,7 +128,7 @@ class NFM(GeneralRecommender):
         pos_pred = self.forward(user, pos_item)
 
         if self.loss_type.upper() in ['CL', 'SL']:
-            label = batch[2].to(self.device)
+            label = batch[2].to(self.device).float()
             loss = self.criterion(pos_pred, label)
 
             loss += self.reg_1 * (self.embed_item(pos_item).norm(p=1))
